added "Clap for Time" functionality. & small thngs
small things:  & => && in Gradient.cpp
added “return 1;” to functions which are supposed to return an int.

+++ LightShowCmd.h
@@ -75,7 +75,7 @@ protected:
 	PixelColor currPixelColor;
 	
 	boolean isFirstTimeAfterFinished;
-	int cmdFinishedTime;
+	unsigned long cmdFinishedTime;
 	
 public:
 	boolean isPauseMode;
